refactor : added a Parsing.ts file for all parsing / string-manipulation related functions (#11)
