+++ api/WikipediaSource.ts
@@ -1,12 +1,4 @@
-import {
-    splitIntoEqualSentenceParts,
-    removeNameOccurrences,
-    getInitials,
-    getAndPermutations,
-    removeTag,
-    months,
-    countries
-} from "~/utilities/Utils";
+import { parseWikitext, splitIntoEqualSentenceParts, removeNameOccurrences, getInitials } from "~/api/Parsing";
 import { type Hint, hintListFromObject } from "~/model/Hint"
 
 /**
@@ -133,113 +125,7 @@ export async function fetchInfoBox(celebrityName: string): Promise<Hint[]> {
         })
 }
 
-/**
- * Private function to parse the wikitext of the first section of a Wikipedia page.
- * It returns an object containing all fields of the infobox.
- * @param wikitext the wikitext of the first section of a Wikipedia page
- */
-function parseWikitext(wikitext: string): {[key: string]: string} {
-    console.log("---------- BEFORE -------------");
-    console.log(wikitext);
-    // Remove introduction after infobox
-    wikitext = wikitext.slice(0, wikitext.toUpperCase().indexOf("'''"));
 
-    // Remove comments and remaining tags
-    wikitext = removeTag(wikitext, "<!--", "-->");
-    wikitext = removeTag(wikitext, "<ref", "</ref>", "/>");
-
-    let hints: {[key: string]: string } = {};
-
-    // Birthdate
-    let match = fieldMatchers.birthDate.exec(wikitext);
-    if (match) {
-        const [, birthYear, birthMonth, birthDay] = match;
-        hints["Born"] = `${birthDay} ${months[parseInt(birthMonth)]} ${birthYear}`;
-    }
-
-    // Death date
-    match = fieldMatchers.deathDate.exec(wikitext);
-    if (match) {
-        const [, deathYear, deathMonth, deathDay] = match;
-        hints["Died"] = `${deathDay} ${months[parseInt(deathMonth)]} ${deathYear}`;
-    }else {
-        hints["Status"] = "Alive";
-    }
-
-    // Description
-    match = fieldMatchers.description.exec(wikitext);
-    let description: string | undefined = match ? match[1] : undefined;
-    if (description) {
-        const {citizenship, occupation} = parseDescription(description);
-        if (citizenship !== undefined) hints["Citizenship"] = citizenship;
-        if (occupation !== undefined) hints["Occupation"] = occupation;
-    }
-
-    // Spouses
-    let spouses: string[] | undefined = undefined;
-    while ((match = fieldMatchers.spouses.exec(wikitext)) !== null) {
-        spouses = spouses ? [...spouses, match[1]] : [match[1]];
-        hints["Spouses"] = spouses.join(', ');
-    }
-
-    // Genres
-    match = fieldMatchers.genres.exec(wikitext);
-    if (match){
-        let genre: string = match[0];
-        let genres: string[] | undefined = undefined;
-        while ((match = fieldMatchers.lists.exec(genre)) !== null) {
-            genres = genres ?
-                [...genres, match[1].split('|')[0].trim()] :
-                [match[1].split('|')[0].trim()];
-            hints["Genres"] = genres.join(', ');
-        }
-    }
-
-    return hints;
-}
-
-function parseDescription(description : string): {citizenship: string | undefined, occupation: string | undefined} {
-
-    let res: {citizenship: string | undefined, occupation: string | undefined} = {
-        citizenship: undefined,
-        occupation: undefined
-    };
-
-    // Check if nationality or country is contained in the description
-    for (const country in countries) {
-        const nationality: string = countries[country];
-        let copy: string = description;
-        while (copy.includes(nationality) || copy.includes(country)) {
-            res.citizenship = res.citizenship !== undefined ? res.citizenship + " and " + nationality : nationality;
-            copy = copy.replace(nationality, "").replace(country, "");
-        }
-    }
-
-    // Check if description matches a known occupation
-    for (const possibleOccupation in occupationMatchers){
-        const matcher: RegExp = occupationMatchers[possibleOccupation];
-        if(description.match(matcher)){
-            res.occupation = possibleOccupation;
-            return res;
-        }
-    }
-
-    // If no occupation was found, occupation is description without citizenship
-    res.occupation = description;
-    for (const permutation of getAndPermutations(res.citizenship || "")){
-        res.occupation = res.occupation.replace(permutation, "").trim();
-    }
-
-    // Handle the `-born`case (i.e. Albert Einstein)
-    if (res.citizenship &&
-        description.includes("-born") &&
-        (description.indexOf("-born") - description.indexOf(res.citizenship) === res.citizenship.length)) {
-        res.citizenship = res.citizenship + '-born';
-        res.occupation = res.occupation.replace("-born", "").trim();
-    }
-
-    return res
-}
 
 
 const BASE_URL: string = "https://en.wikipedia.org";
