+++ utilities/Utils.ts
@@ -20,117 +20,6 @@ export function getRandom(list: any[]): any {
 }
 
 /**
- * Given a name (string) this function returns a string containing the initials of the name
- * @param name - name we want the initials of
- */
-export function getInitials(name: string): string {
-    // Split the name into individual words
-    const words: string[] = name.split(' ');
-
-    // Initialize an empty string to store the initials
-    let initials: string = '';
-
-    // Iterate through each word and add its first letter to the initials
-    for (const word of words) {
-        // If the word is not empty, add its first letter to the initials
-        if (word.length > 0) {
-            initials += word[0].toUpperCase() + ". ";
-        }
-    }
-
-    return initials;
-}
-
-/**
- * Given a sentence of words separated by "and", returns all
- * possible permutations of the words around these "and".
- * @param input - a sentence of words separated by "and"
- */
-export function getAndPermutations(input: string): string[] {
-    const words = input.split(" and ").map(word => word.trim());
-
-    function generatePermutations(arr: string[], index: number) {
-        if (index === arr.length - 1) {
-            return [arr.slice()];
-        }
-
-        const result: string[][] = [];
-        for (let i = index; i < arr.length; i++) {
-            [arr[index], arr[i]] = [arr[i], arr[index]]; // Swap
-            result.push(...generatePermutations(arr, index + 1));
-            [arr[index], arr[i]] = [arr[i], arr[index]]; // Restore
-        }
-        return result;
-    }
-
-    const permutations: string[] = [];
-    const permutedArrays = generatePermutations(words, 0);
-    for (const permutedArray of permutedArrays) {
-        permutations.push(permutedArray.join(" and "));
-    }
-
-    return permutations;
-}
-
-/**
- * Remove all text  between the opening and closing tags in the given string.
- * @param text - a string that may contain tags to remove
- * @param opening - the opening tag
- * @param closings - the closing tag
- */
-export function removeTag(text: string, opening: string, ...closings: string): string {
-    const regexStr = `${opening}[^]*?(${closings.join('|')})`;
-    const regex = new RegExp(regexStr, 'g');
-    return text.replace(regex, '');
-}
-
-/**
- * Remove all occurrences of the given name in the given text, including
- * first name and last name, excluding name particles (e.g. "von", "de", ...).
- * @param name - celebrity first name(s) and last name(s)
- * @param text - text from which to remove the name
- */
-export function removeNameOccurrences(text: string, name: string): string {
-    // Get lists of all first names and names, without name particle (i.e. von, de, of...)
-    const names: string[] = [
-        ...name.split(" ").map(n => n.trim()).filter(n => n === capitalize(n)), name
-    ]
-    names.forEach(n => {
-        const regex = new RegExp(n, 'gi');
-        text = text.replace(regex, "???");
-    })
-    return text
-}
-
-/**
- * Split the given text into parts containing equal number of sentences.
- * @param text - the text to split
- * @param num - the number of parts to split the text into
- */
-export function splitIntoEqualSentenceParts(text: string, num: number): string[] {
-    // Split the text into sentences
-    const sentences = text.match(/[^.!?]+[.!?]+/g);
-    if (!sentences) return [text];
-
-    const totalSentences = sentences.length;
-    const sentencesPerPart = Math.floor(totalSentences / num); // Number of sentences per part
-    const parts: string[] = [];
-
-    // Distribute sentences evenly across the three parts
-    let start = 0;
-    for (let i = 0; i < num - 1; i++) {
-        const end = start + sentencesPerPart;
-        parts.push(sentences.slice(start, end).join(''));
-        start = end;
-    }
-
-    // Add the remaining sentences to the last part
-    parts.push(sentences.slice(start).join(''));
-
-    return parts;
-}
-
-/**
  * Get random user model
  * @returns a random user model
  */
